movement of the blocks to appropriate modules.
